test: refactor Document-integ-test
